+++ src/models/coureur_point.rs
@@ -1,5 +1,4 @@
 use async_graphql::SimpleObject;
-use bigdecimal::BigDecimal;
 use diesel::prelude::*;
 
 use crate::view::v_coureur_point;
@@ -12,6 +11,6 @@ use crate::view::v_coureur_point;
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct CoueurPoint {
     pub coureur: i32,
-    pub points: BigDecimal,
-    pub temps: BigDecimal,
+    pub points: i64,
+    pub temps: i64,
 }
